Fix decoding of compressed NOP instructions
C.ADDI with an immediate=0 is normally not allowed, except when rd=x0
when it is used to encode a NOP instruction

See Section 16.5 of  “The RISC-V Instruction Set Manual, Volume I: User-Level ISA, Document Version20191213”, Editors Andrew Waterman and Krste Asanovic, RISC-V Foundation, December 2019.

Thanks to Lukas Döllerer for pointing this out.

Work done as part of eragp-sbt-2021 .